+++ lib/decidim/decidim_awesome/test/shared_examples/editor_examples.rb
@@ -2,7 +2,7 @@
 
 shared_examples "has no image support" do
   it "has no image button" do
-    expect(page).to have_no_xpath("//button[@class='editor-toolbar-control'][@data-editor-type='image']")
+    expect(page).not_to have_xpath("//button[@class='editor-toolbar-control'][@data-editor-type='image']")
   end
 end
 
@@ -14,7 +14,7 @@ end
 
 shared_examples "has no video support" do
   it "has no video button" do
-    expect(page).to have_no_xpath("//button[@class='editor-toolbar-control'][@data-editor-type='videoEmbed']")
+    expect(page).not_to have_xpath("//button[@class='editor-toolbar-control'][@data-editor-type='videoEmbed']")
   end
 end
 
@@ -26,7 +26,7 @@ end
 
 shared_examples "has no drag and drop" do
   it "cannot drop a file" do
-    expect(page).to have_no_content("Add images by dragging & dropping or pasting them.")
+    expect(page).not_to have_content("Add images by dragging & dropping or pasting them.")
     find(editor_selector).drop(image)
     expect(page.execute_script("return document.querySelector('#{editor_selector}').value")).not_to eq("[Uploading file...]")
     sleep 1
@@ -37,19 +37,23 @@ end
 
 shared_examples "has drag and drop" do
   it "can drop a file" do
-    expect(page).to have_content("Add images by dragging & dropping or pasting them.")
-    fill_in "proposal_title", with: "This is a test proposal"
-    fill_in "proposal_body", with: ""
+    within "form.edit_proposal" do
+      expect(page).to have_content("Add images by dragging & dropping or pasting them.")
+      fill_in :proposal_body, with: ""
+    end
     find(editor_selector).drop(image)
-    expect(page.execute_script("return document.querySelector('#{editor_selector}').value")).to eq("[Uploading file...]")
+    expect(page.execute_script("return document.querySelector('#{editor_selector}').value")).to include("[Uploading file...]")
     sleep 1
     last_image = Decidim::DecidimAwesome::EditorImage.last
     expect(last_image).to be_present
     content = page.execute_script("return document.querySelector('#{editor_selector}').value")
     expect(content).to include(last_image.attached_uploader(:file).path)
-    # ensures valid body validations
-    fill_in "proposal_body", with: "This is a super test\n#{content}"
-    click_link_or_button "Send"
+    within "form.edit_proposal" do
+      # ensures valid body validations
+      fill_in :proposal_body, with: "This is a super test with lots of downcases characters to be sure that the image name does not mess with percentage of caps.\n#{content}"
+      fill_in :proposal_title, with: "This is a test proposal"
+      click_on "Send"
+    end
     expect(proposal.reload.body["en"]).to include(last_image.attached_uploader(:file).path)
   end
 end